refactor: move server-side modules into src/server
