--- src/actions/authorize.js
@@ -1,1 +1,0 @@
-// permissions for pages
