+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,2 @@
-import { handlers } from "@/authentication";
+import { handlers } from "@/server/auth/config";
 export const { GET, POST } = handlers;
