+++ src/app/layout.js
@@ -1,7 +1,7 @@
 import "@ant-design/v5-patch-for-react-19";
 import "antd/dist/reset.css";
 
-import { auth } from "@/authentication";
+import { auth } from "@/server/auth/config";
 
 import "./globals.css";
